generate: Drop /dev/mqueue
Linux runtimes are not required to supply this [1,2].  We could supply
it via ocitool, but I don't think its wise to try and define two
standards (one in opencontainers/specs for what a runtime must supply,
and another here with additional stuff), because I don't see a clear
definition of something generally useful enough to go here but not in
the spec.  For example, hosts without CONFIG_POSIX_MQUEUE could be
spec-compliant but unable to perform the mqueue mount I'm removing
here.

[1]: https://github.com/opencontainers/specs/blob/v0.1.1/config-linux.md#default-devices-and-file-systems
[2]: https://github.com/opencontainers/specs/pull/164

Signed-off-by: W. Trevor King <wking@tremily.us>